// configs/props/interface.ts
interface LocationInfoProps {
    coordinates: {
        lat: number,
        lng: number,
    },
    name: string,
    photoRef: string,
    url: string
}

interface TravellerProps {
    desc: string,
    icon: any,
    id: number,
    people: string,
    title: string
}

export interface TripDataProps {
    locationInfo: LocationInfoProps,
    startDate: string,
    endDate: string,
    totalNoOfDays: number,
    traveller: TravellerProps,
    budget: string
}

export interface FlightProps {
    bookingUrl: string,
    details: string,
    price: string
}

interface geoCoordinatesProps {
    latitude: string,
    longitude: string,
}

export interface HotelDetailsProps {
    address: string,
    description: string,
    geoCoordinates: geoCoordinatesProps,
    hotelName: string,
    imageUrl: string,
    price: string,
    rating: string
}

export interface ItineraryPlanProps {
    geoCoordinates: geoCoordinatesProps,
    placeDetails: string,
    placeImageUrl: string,
    placeName: string,
    ticketPricing: string,
    timeToTravel: string
}

export interface ItineraryProps {
    day: number,
    plans: ItineraryPlanProps[],
    location: string,
    travelers: string
}

interface TravelPlanProps {
    budget: string,
    duration: string,
    flight: FlightProps,
    hotels: HotelDetailsProps[],
    itinerary: ItineraryProps[]
}

interface TripPlanProps {
    travelPlan: TravelPlanProps
}

// data record
export interface UserTripsProps {
    docId: string,
    userEmail: string,
    tripData: string,    // TripDataProps as string to save space; input for prompt
    tripPlan: TripPlanProps // generated by AI(Gemini or OpenAI)
}